fix: extension attach content from project input bar
