+++ extension/ui/pages/ProjectMainPage.tsx
@@ -1,8 +1,8 @@
-import { InputBarProvider } from "@app/components/assistant/conversation/input_bar/InputBarContext";
 import { SpaceConversationsPage } from "@app/components/pages/conversation/SpaceConversationsPage";
 import { useAuth } from "@app/lib/auth/AuthContext";
 import { useSpaceInfo } from "@app/lib/swr/spaces";
 import { ConversationLayout } from "@extension/ui/components/conversation/ConversationLayout";
+import { ExtensionInputBarProvider } from "@extension/ui/components/conversation/ExtensionInputBarProvider";
 import { useParams } from "react-router-dom";
 
 export const ProjectMainPage = () => {
@@ -16,9 +16,9 @@ export const ProjectMainPage = () => {
 
   return (
     <ConversationLayout title={spaceInfo?.name ?? ""}>
-      <InputBarProvider>
+      <ExtensionInputBarProvider workspace={workspace}>
         <SpaceConversationsPage />
-      </InputBarProvider>
+      </ExtensionInputBarProvider>
     </ConversationLayout>
   );
 };
